Specify day of the week as day name
Switch from the internal Floodgate format for specifying the
day of the week on which the package upgrade image is built (numbers 0-6
as strings), to using the day names (e.g. "Monday") as values for the
parameter `imageBuildDay` (renamed from `imageDay`).

This makes the configuration clearer at the expense of needing a
translation layer to the internal Floodgate format.